chore: valeurComptableFutur return type to float

+++ src/main/java/school/hei/patrimoine/possession/Argent.java
@@ -10,7 +10,7 @@ public final class Argent extends Possession {
   }
 
   @Override
-  public int valeurComptableFuture(Instant tFutur) {
+  public float valeurComptableFuture(Instant tFutur) {
     throw new NotImplemented();
   }
 }
